Author: Isomer <isomer@coders.net> Log message:
Fixed lots of compiletime warnings.  Thanks to CyberDude that gave me
a shell so I could test.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@269 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ ircd/s_user.c
@@ -900,7 +900,9 @@ static unsigned char hash_target(unsigned int target)
  */
 void add_target(struct Client *sptr, void *target)
 {
-  unsigned char  hash = hash_target((unsigned int) target);
+  /* Ok, this shouldn't work esp on alpha
+  */
+  unsigned char  hash = hash_target((unsigned long) target);
   unsigned char* targets;
   int            i;
   assert(0 != sptr);
@@ -933,7 +935,7 @@ void add_target(struct Client *sptr, void *target)
 int check_target_limit(struct Client *sptr, void *target, const char *name,
     int created)
 {
-  unsigned char hash = hash_target((unsigned int) target);
+  unsigned char hash = hash_target((unsigned long) target);
   int            i;
   unsigned char* targets;
